Added additional unit tests for NXclassWriters
